LayoutTests/http/tests/security/contentSecurityPolicy/combine-multiple-policies-expected.txt

CONSOLE MESSAGE: Refused to execute inline script because of Content-Security-Policy.

CONSOLE MESSAGE: Refused to apply inline style because of Content-Security-Policy.

This test checks that we enforce all the supplied policies. This test passes if it doesn't alert fail and if the style doesn't apply.
